Bug fixes. Light positioning takes hierarchy into account now.

+++ Firefly/World/Scene/SceneDataModels/Camera.cs
@@ -8,10 +8,10 @@ namespace Firefly.World.Scene.SceneDataModels
   {
     public int projectionType { get; set; }
     public float fov { get; set; }
-    public int farClip { get; set; }
+    public float farClip { get; set; }
     public float nearClip { get; set; }
-    public List<int> position { get; set; }
-    public List<int> rotation { get; set; }
+    public List<float> position { get; set; }
+    public List<float> rotation { get; set; }
     public string skybox { get; set; }
   }
 }
